requirements.txt
Flask==2.0.2
Flask-JWT-Extended==4.3.1
